+++ app/modules/editor/components/StaticFormInputs/TableBaseInput.tsx
@@ -2,8 +2,8 @@ import React, { FC } from 'react';
 import styled from 'styled-components/native';
 import LongTouchButton from '../../../../components/LongTouchButton';
 import { useAppDispatch } from '../../../../redux/hooks';
-import { useEditorBaseSelector } from '../../../../modules/editor/redux/editorSelectors';
-import { changeTableBase } from '../../../../modules/editor/redux/actions/composed/changeTableBase';
+import { useEditorBaseSelector } from '../../redux/editorSelectors';
+import { changeTableBase } from '../../redux/actions/composed/changeTableBase';
 
 export const ButtonsSet = styled.View`
   flex: 1;
