+++ app/modules/editor/components/StaticFormInputs/TableTypeInput.tsx
@@ -1,10 +1,10 @@
 import React, { FC } from 'react';
 import styled from 'styled-components/native';
 import LongTouchButton from '../../../../components/LongTouchButton';
-import { TableTypeEnum } from '../../../../modules/editor/enums';
+import { TableTypeEnum } from '../../enums';
 import { useAppDispatch } from '../../../../redux/hooks';
-import { useEditorBaseSelector, useEditorTypeSelector } from '../../../../modules/editor/redux/editorSelectors';
-import { changeTableType } from '../../../../modules/editor/redux/actions/composed/changeTableType';
+import { useEditorBaseSelector, useEditorTypeSelector } from '../../redux/editorSelectors';
+import { changeTableType } from '../../redux/actions/composed/changeTableType';
 
 /**
  * STYLES
