+++ app/modules/editor/editorTypes.ts
@@ -2,11 +2,12 @@ import { ImmutableJSType } from '../../redux/types';
 import { TableTypeEnum, SetTypeEnum } from './enums';
 
 export type EditorStateType = {
-  trainingTable: TableType;
+  trainingTable: TrainingTableType;
   sets: TableSetType[];
 };
+export type ImmutableJSEditorStateType = EditorStateType & ImmutableJSType;
 
-export type TableType = {
+export type TrainingTableType = {
   base: number;
   type: TableTypeEnum;
   duration: number;
@@ -14,15 +15,13 @@ export type TableType = {
   baseBreaks?: number;
   enduranceLaps?: number;
 };
+export type ImmutableTrainingTableType = TrainingTableType & ImmutableJSType;
 
 export type TableSetType = {
   duration: number;
   type: SetTypeEnum;
   pos: number;
 };
+export type ImmutableJSEditorSetType = TableSetType & ImmutableJSType;
 
 // TODO: Removing immubtale
-export type ImmutableJSEditorStateType = ImmutableJSType;
-export type ImmutableJSEditorType = ImmutableJSType;
-export type ImmutableTrainingTableType = ImmutableJSType;
-export type ImmutableJSEditorSetType = TableSetType & ImmutableJSType;
